requirements.txt
fastapi
uvicorn
transformers
torch
pymongo
sentencepiece
streamlit
pandas
python-dotenv
pyperclip
